requirements.txt
#default
django==2.2.18.0
dj-database-url==0.5.0
django-storage-url==0.5.0
whitenoise==5.2.0
boto3==1.14.49
psycopg2==2.8.5
uwsgi==2.0.19.1
django-cms==3.8.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-history==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-snippet==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==3.9.0
djangocms-video==3.0.0
django-filer==2.0.2
django-parler==2.2
django-select2==7.7.1
# django-shop
django-treebeard==4.4 # fix tree bug
django-shop==1.2.3
django-filter==2.2.0
django-ipware==3.0.1
django-sass-processor==0.8
djangocms-cascade==1.3.7
libsass==0.19.4
# dmanage
openpyxl==3.0.5
# vuejs
django-webpack-loader==0.6.0
# payment providers
stripe==2.55.0
squareup==6.3.0.20200826
# extras
django-autocomplete-light==3.8.1
django-admin-ordering==0.13.1
django-autoslug==1.9.8
django-colorfield==0.3.0
django-haystack==3.0
geopy==2.1.0
mailchimp3==3.0.7
python-slugify==4.0.1
# tools
coverage==5.4
django-mptt==0.11.0
bs4==0.0.1
django-cities-light==3.8.1
sentry-sdk
lxml

